Restore mode without spatial locality
If no spatial proximity edges are present in the dependence graph, use
old scheduling approach.
The LP is still slightly different due to a sharing artifact: dimensions
for distance bounding function parameters start from 5 and 6 rather than
from 4 and 5 and dimension 4 is reserved for the constant of that
function (essentially equal to dimension 1).