test_inputs/codegen/{atomic,separate}.*: explicitly specify order of statements
For these test cases, the same output is used for the union map input and
the schedule tree input.  In the original input, the order of
the statements is not specified, so the order in the output
depends on the hash values used in the union map representation
in a largely unpredictable way.  The two outputs could therefore
be different.  While they happens to be the same at the moment,
they may start to differ when changes are made to the way
hash values are computed.
Specify the order of the statements to ensure that the output
is always the same.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>