isl_scheduler.c: intra_coefficients: separate cache for parametric constraints
In an upcoming commit, the isl scheduler will remove some constraints
on the schedule coefficients that could be exploited for constructing
loop coalescing schedules.  This removal is most effective if
there are no parameter coefficients.  The parameters will therefore
be projected out first.
However, this only works when the constraints on schedule coefficients
are used to enforce non-negativity and not when they are used
for bounding the dependence distance.
In particular, bounding the dependence distance requires
parameter coefficients and may also be able to favorably
use the constraints that are removed in the other case.

The construction of the constraints on schedule coefficients
will therefore be different for the two use cases and
they therefore need to be stored in separate caches.
Introduce a separate cache for those that require
parameter coefficients.

The intra_coefficients is given an extra parameter indicating
whether the caller requires parameter coefficients.
When the caller does not require them, intra_coefficients may
still include them in the result.
This reduces the number of cases intra_coefficients
needs to take into account when the constraint removal
will be introduced.
Note that the callers that do not require parameter coefficients
will plug in zero for these parameter coefficients anyway.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>