isl_scheduler.c: collect_intra_validity: drop parameters
In the next commit, the isl scheduler will remove some constraints
on the schedule coefficients that could be exploited for constructing
loop coalescing schedules.  This removal is most effective if
there are no parameter coefficients.
Remove the parameters in collect_intra_validity.

Note that collect_intra_validity does not need
parameter coefficients since it would plug in zero for them anyway.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>