isl_sched_graph: add identifier to compressed nodes for differentiation
If any node is compressed, then the set for which coefficients
are computed lies in the compressed space.
These spaces are currently only identified by their dimension,
meaning that multiple nodes may have sets in the same compressed space.
In an upcoming commit, such sets or related binary relations
will be grouped together and they therefore need to be made to live
in different spaces.
Add a name and a unique user pointer to the compressed spaces
for this purpose.
The unique user pointer is taken to be a pointer to the node itself.
This allows the node to be easily recovered from the space.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>