isl_sched_graph: ignore parameters in identification of nodes
The nodes in the graph were hashed on the complete space,
include parameters.  This means that the parameters of a space
passed to graph_find_node need to be aligned with those of the spaces
used during the construction of the graph.
In an upcoming commit, graph_find_node may get called on a space
derived from a set without parameters, in particular one derived
from the space of coefficients.
While it is possible to add the parameters back in,
it is easier to simply ignore the parameters in the space
since they are not needed to identify a node anyway.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>