scheduler: exploit lineality space in difference set of intra-node dependences
If the recently introduced mechanism for removing constraints
that can only be used for coalescing kicks in, then
after this removal, the difference set of intra-node
validity dependences may have coordinates that are completely
unconstrained.  This means that the corresponding schedule coefficients
are required to be zero.  This information can be exploited
beforehand by extending all sets of dependences to sets on
equivalent elements.  This may in particular cause some disjuncts
to become identical, reducing the total number of disjuncts
that need to be considered for carrying.

For example, if the difference set of schedule constraints
(after removal of coalescing constraints) is of the form

{ [a,b,c] : a > 0 or (a = 0 and b > 0) or (a = 0 and b = 0 and c > 0) }

then the combined lineality space is { { [a,b,c] : a = 0 }.
Extending the difference set to elements with the same value for a
results in

{ [a,b,c] : a > 0 or (a = 0) or (a = 0) }

where the last two disjuncts can be replaced by a single disjunct.

The detection of non-trivial lineality spaces is currently
applied unconditionally, but it is unlikely to find anything
if no coalescing constraints have been removed.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>
Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>