isl_scheduler.c: compute_schedule_finish_band: also carry coincidence edges
In particular, if the Pluto-like scheduler did not find a solution and
if the schedule_outer_coincidence option is set,
then the failure to find a solution may be due to the coincidence
constraints.
Theoretically, there may be no validity constraints, but only
coincidence constraints, in which case some of those coincidence
constraints should get carried in order to make progress.
Include the coincidence constraints under these circumstances.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>