imath backend: use tailored isl_int_fdiv_q_ui
In particular, simply convert the unsigned long argument to an mpz_t and
call impz_fdiv_q.  impz_fdiv_q_ui does essentially the same, but it
computes the integer division twice, once to get the quotient and
once to get the remainder.  The remainder is needed because
impz_fdiv_q_ui is meant to mimic gmp's mpz_fdiv_q_ui, which
returns this remainder.  However, there is no need for
the isl_int_fdiv_q_ui macro to return this remainder.

This change is unlikely to result in any measurable performance
improvement, but it still feels better to only compute the integer
division once.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>