isl_hash_{space,space_domain}: extract out shared parameter hashing
isl_space_get_hash takes into account both the parameters and
the tuples.  In some cases, it can by convenient to be able
to ignore the parameters.  A new function isl_space_get_tuple_hash
will be introduced for this purpose, which will (conceptually)
share the handling of the tuples with isl_space_get_hash.
In order to also be able to share the code for handling tuples,
it first needs to be separated from the handling of parameters.
This separation is performed by this commit,
which also removes some code duplication.

Since the handling or parameters and tuples was intermixed,
the separation results in changes in the hash values and
subsequently in changes in the AST generation output.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>