isl_space_get_{hash,domain_hash}: only hash in parameters once
isl_hash_space and isl_hash_space_domain would call themselves
recursively on nested spaces, resulting in the (same) parameters
needlessly being hashed in twice.
Extract out the hashing in of parameters such that it is only
performed once (in the caller).  The names of the functions are
changed to reflect this change in behavior.
After removing the handling of parameters, it will also be possible
to reuse the resulting isl_hash_tuples in the function
isl_space_get_tuple_hash introduced in the next commit.

Since the parameters are only hashed in once, the hash values
may be different, which results in some changes in the AST generation output.

Signed-off-by: Sven Verdoolaege <skimo@kotnet.org>