isl_scheduler.c: construct_trivial: construct region over all node coefficients
Each non-triviality region was only defined over the LP variables
corresponding to the last schedule coefficients.
In an upcoming commit, the order of the coefficients will be changed
such that the triviality directions no longer correspond to
the last variables.
Define each non-triviality region over all variables
corresponding to the schedule coefficients of a node.
This will make it easier to handle different relations
between the coefficients and the triviality directions.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>