ignore computed schedule of timing test case
Commit b57971ad29 (isl_tab_basic_set_non_trivial_lexmin: do not add cuts
for all variables, Sat Feb 18 16:13:23 2012 +0100) introduced a test case
that would result in extremely long computation times without
the changes in that commit.  The actual form of the resulting schedule
was not the object of the test.
Change the test such that it only checks that a schedule can be computed
(within a reasonable amount of time) rather than checking that is equal to
some specific schedule.
The next commit results in another schedule being computed
for this test case and the change in output would be a mere distraction.

Note that as a side-effect of commit 39d7b91232 (isl_hash_dim: make result
independent of endianness, Fri Dec 7 17:51:13 2012 +0100), the order
of some internal objects has changed such that in the end this test
case would no longer result in extremely long computation times even
without the changes in the commit that introduced the test case.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>