add isl_schedule_constraints_set_intra_consecutivity
These constraints allow the user to specify preferential choices
for the inner schedule rows of one or more domain spaces.
In particular, each constraint in the list specifies
that the outer schedule rows should be linear combinations
of some rows, while the inner schedule rows should be equal
(modulo linear combinations of outer schedule rows) to
another sequence of rows.

The typical use case is to achieve consecutive accesses
in the innermost index expression of an array access.
In this case, the outer schedule rows should be linear
combinations of the outer index expressions, while
the innermost schedule row should be equal to the innermost
index expression.

If there are multiple constraints per domain space,
then, at every given schedule level, they are considered in turn
until progress can be made on one of them.
All those that fail to make progress are no longer considered
at subsequent levels.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>