Complete isl_aff to set interface
Document and export the function isl_aff_gt_set. There is no obvious difference
to all the existing isl_aff and isl_pw_aff funtions that perform the very
same operation with a different comparision operator. To complete the
interface we export the missing function.

Implement, document, and export the function isl_aff_ne_set. The
function isl_aff_ne_set is special in comparision to the other operands,
as for this function it is not possible to construct a basic_set,
because a not-equal comparision my result in two disjuncts. Hence, only
an isl_aff_ne_set version can be introduced, but no isl_aff_gt_basic_set
version. To complete the interface, the isl_aff_ne_set is introduced.

Signed-off-by: Tobias Grosser <tobias@grosser.es>