interface/cpp.cc: store C++ callback function in a struct
Store the C++ callback function in a struct for passing
to the C callback function.  This avoids the need for
a (naked) double indirection on the C++ callback function and
should be less error-prone because the type of the callback
function is now checked by the compiler.
The use of a struct also allows additional pieces
of data to be passed to the C callback function.
This will be exploited in the next commit.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>