cpp: generate isl::managed_copy for all exported classes
Provide a set of overloaded functions isl::managed_copy(__isl_keep isl_class*)
that return a C++-wrapped copy of the given object. Contrary to isl::manage,
these functions do not take ownership of the given object and will not destroy
it when the last C++ copy goes out of scope.  This simplifies the interaction
with C code using isl, specifically with other libraries such as ppcg.

Signed-off-by: Oleksandr Zinenko <git@ozinenko.com>