Add isl_space_[set|map|params]_alloc_from_id_list
This allows the allocation of an isl_space with a list of parameter ids.
As a result, the space that is returned contains parameter dimensions
that are uniquely identified by an isl_id. Such parameter dimensions are
also called named parameter dimensions.

Named parameter dimensions are uniquely identified by the isl_id they
reference and compare according to the rules for isl_ids (see manual):

  "Identifiers with the same name but different pointer values are
   considered to be distinct. Similarly, identifiers with different names
   but the same pointer value are also considered to be distinct. Equal
   identifiers are represented using the same object."

Also add isl_space_[set|map]_noparams and isl_space_params_alloc_empty
for the allocation of spaces without any parameters. These will later
serve as convenience alternatives in the python/c++ interfaces, where
the allocation functions that allow for the creation of unnamed
parameter dimensions are not available.

Signed-off-by: Tobias Grosser <tobias@grosser.es>