export more isl_schedule_node functionality
Functions that are not exported are those that require support
for extra types or features such as isl_id, isl_space and callbacks
that return anything other than isl_stat.
The function isl_schedule_node_foreach_ancestor_top_down is not exported
because, unlike other exported functions that return isl_stat,
it does not take its argument.
The function isl_schedule_node_band_sink is not exported because
the generator assumes that an object returned by a subclass
method is of the subclass type, which may not be the case
for isl_schedule_node_band_sink.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>