interface: add support for type based subclasses
Some isl types such as isl_ast_node and isl_schedule_node
represent a collection of subclasses that each have their
own corresponding functions.  In C, the user is responsible
for checking the type of the object before calling
the appropriate functions.

In the C++ or Python bindings, the subclasses can instead
be made available as proper subclasses.
Add support for generating these subclasses from
the function that returns the type of the subclass.
In particular, any function with an __isl_subclass annotation
is treated as such a type function.

The generated code is slightly different for C++ and Python.
Since Python is a dynamically-typed language, each object
returned by a C function can be cast directly to the appropriate
subclass.  For this purpose, a __new__ method is introduced
that constructs a Python object of the subclass that corresponds
to the type of the C object.  The __new__ method is overridden
in the subclasses in order to be able to make regular objects
of these subclasses.
The constructor of the superclass is also extended to accept
objects of the subclasses in order to support the general casting
that is performed when an object is not of the expected type.

In the C++ bindings, the (static) type of an object returned by
a superclass method is that of the superclass.  This superclass
has two methods for checking the dynamic type and for downcasting
to this type, in particular

	bool super::isa<sub>();
and
	sub super::as<sub>();

The as() method returns an invalid object if the concrete object
is not of the expected type.  This allows a construct of the form

        if (auto filter = node.as<isl::schedule_node_filter>()) {
            /* operations of filter node */
        }

The type of an object returned by a subclass method is assumed
to be that of the subclass.  This assumption may not apply
to all functions that could potentially be exported.  Further
changes may therefore be required before these functions
can be exported.

Signed-off-by: Sven Verdoolaege <sven.verdoolaege@gmail.com>

update return type